Dried JS code, fixed several bugs, added comments
Made JS more DRY. Fixed bugs handling initial temps of zero with final temps of 100. Fixed heat of fusion and vaporization always being included in the result. Fixed off by one errors due to "<" needing to be "<=" in calculating differentials dependent on the initial temperature at the heat of fusion/vaporization. Added empty mass and not a number exceptions for the input fields. Added comments to make sense of abbreviations.